refactor(c3)!: drop const in devices
