+++ midealocal/devices/c3/__init__.py
@@ -2,13 +2,15 @@
 
 import json
 import logging
+from enum import StrEnum
 from typing import Any, ClassVar, Unpack
 
 from midealocal.const import DeviceType
 from midealocal.device import MideaDevice, MideaDeviceInitKwargs
 
-from .const import C3DeviceMode, C3SilentLevel, DeviceAttributes
 from .message import (
+    C3DeviceMode,
+    C3SilentLevel,
     MessageC3Response,
     MessageQuery,
     MessageQueryBasic,
@@ -23,6 +25,55 @@ from .message import (
 )
 
 _LOGGER = logging.getLogger(__name__)
+
+
+class DeviceAttributes(StrEnum):
+    """Midea C3 device attributes."""
+
+    zone1_power = "zone1_power"
+    zone2_power = "zone2_power"
+    dhw_power = "dhw_power"
+    zone1_curve = "zone1_curve"
+    zone2_curve = "zone2_curve"
+    disinfect = "disinfect"
+    fast_dhw = "fast_dhw"
+    zone_temp_type = "zone_temp_type"
+    zone1_room_temp_mode = "zone1_room_temp_mode"
+    zone2_room_temp_mode = "zone2_room_temp_mode"
+    zone1_water_temp_mode = "zone1_water_temp_mode"
+    zone2_water_temp_mode = "zone2_water_temp_mode"
+    mode = "mode"
+    mode_auto = "mode_auto"
+    zone_target_temp = "zone_target_temp"
+    dhw_target_temp = "dhw_target_temp"
+    room_target_temp = "room_target_temp"
+    zone_heating_temp_max = "zone_heating_temp_max"
+    zone_heating_temp_min = "zone_heating_temp_min"
+    zone_cooling_temp_max = "zone_cooling_temp_max"
+    zone_cooling_temp_min = "zone_cooling_temp_min"
+    tank_actual_temperature = "tank_actual_temperature"
+    room_temp_max = "room_temp_max"
+    room_temp_min = "room_temp_min"
+    dhw_temp_max = "dhw_temp_max"
+    dhw_temp_min = "dhw_temp_min"
+    target_temperature = "target_temperature"
+    temperature_max = "temperature_max"
+    temperature_min = "temperature_min"
+    status_heating = "status_heating"
+    status_dhw = "status_dhw"
+    status_tbh = "status_tbh"
+    status_ibh = "status_ibh"
+    total_energy_consumption = "total_energy_consumption"
+    total_produced_energy = "total_produced_energy"
+    outdoor_temperature = "outdoor_temperature"
+    temp_tw_in = "temp_tw_in"
+    temp_tw_out = "temp_tw_out"
+    instant_power0 = "instant_power0"
+    silent_mode = "silent_mode"
+    SILENT_LEVEL = "silent_level"
+    eco_mode = "eco_mode"
+    tbh = "tbh"
+    error_code = "error_code"
 
 
 class MideaC3Device(MideaDevice):
@@ -137,27 +188,27 @@ class MideaC3Device(MideaDevice):
                         == C3DeviceMode.COOL
                     ):  # cooling mode
                         self._attributes[DeviceAttributes.temperature_max][zone] = (
-                            self._attributes[
-                                DeviceAttributes.zone_cooling_temp_max
-                            ][zone]
+                            self._attributes[DeviceAttributes.zone_cooling_temp_max][
+                                zone
+                            ]
                         )
                         self._attributes[DeviceAttributes.temperature_min][zone] = (
-                            self._attributes[
-                                DeviceAttributes.zone_cooling_temp_min
-                            ][zone]
+                            self._attributes[DeviceAttributes.zone_cooling_temp_min][
+                                zone
+                            ]
                         )
                     elif (
                         self._attributes[DeviceAttributes.mode] == C3DeviceMode.HEAT
                     ):  # heating mode
                         self._attributes[DeviceAttributes.temperature_max][zone] = (
-                            self._attributes[
-                                DeviceAttributes.zone_heating_temp_max
-                            ][zone]
+                            self._attributes[DeviceAttributes.zone_heating_temp_max][
+                                zone
+                            ]
                         )
                         self._attributes[DeviceAttributes.temperature_min][zone] = (
-                            self._attributes[
-                                DeviceAttributes.zone_heating_temp_min
-                            ][zone]
+                            self._attributes[DeviceAttributes.zone_heating_temp_min][
+                                zone
+                            ]
                         )
                 else:  # Room temp mode
                     self._attributes[DeviceAttributes.target_temperature][zone] = (
